Terraform/lambda_functions/requirements.txt
PyJWT==2.8.0
boto3==1.34.144
requests==2.31.0
cryptography==41.0.7